`createToaster` approach, much cleaner!
