--- packages/skeleton-react/src/components/Toast/context.ts
@@ -1,6 +1,0 @@
-'use client';
-
-import { createContext } from 'react';
-import type { ToastContextType } from './types.js';
-
-export const ToastContext = createContext<ToastContextType>(null!);
