+++ packages/skeleton-react/src/index.ts
@@ -12,8 +12,8 @@ export { Segment } from './components/Segment/Segment.js';
 export { Slider } from './components/Slider/Slider.js';
 export { Switch } from './components/Switch/Switch.js';
 export { Tabs } from './components/Tabs/Tabs.js';
-export { ToastProvider } from './components/Toast/ToastProvider.js';
-export { ToastContext } from './components/Toast/context.js';
+export { Toaster } from './components/Toast/Toaster.js';
+export { createToaster } from './components/Toast/create-toaster.js';
 
 // Types
 export type { FileUploadApi } from './components/FileUpload/types.js';
