--- packages/skeleton-svelte/src/components/Toast/context.ts
@@ -1,4 +1,0 @@
-import { createContext } from '../../internal/create-context.js';
-import type { ToastContext } from './types.js';
-
-export const [setToastContext, getToastContext] = createContext<ToastContext>(null!);
