+++ playgrounds/skeleton-react/src/app/lib/toaster.ts
@@ -1,0 +1,6 @@
+import { createToaster } from '@skeletonlabs/skeleton-react';
+
+export const toaster = createToaster({
+	overlap: true,
+	placement: 'bottom-end'
+});
