+++ playgrounds/skeleton-svelte/src/lib/toaster.ts
@@ -1,0 +1,5 @@
+import { createToaster } from '@skeletonlabs/skeleton-svelte';
+
+export const toaster = createToaster({
+	placement: 'bottom-end'
+});
