+++ playgrounds/skeleton-svelte/src/routes/+layout.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
 	import '../app.css';
 	// Components
-	import { ToastProvider, Switch } from '@skeletonlabs/skeleton-svelte';
+	import { Toaster, Switch } from '@skeletonlabs/skeleton-svelte';
 	// Icons
 	import IconMoon from '@lucide/svelte/icons/moon';
 	import IconSun from '@lucide/svelte/icons/sun';
+	import { toaster } from '../lib/toaster.js';
 
 	let { children } = $props();
 	let modeState = $state(false); // false = dark mode
@@ -14,63 +15,63 @@
 	}
 </script>
 
-<ToastProvider placement="bottom-end">
-	<div class="grid h-screen grid-cols-[320px_minmax(0,_1fr)]">
-		<!-- Nav -->
-		<div class="bg-surface-100-900 space-y-8 overflow-y-auto p-8">
-			<header>
-				<a class="text-sm bg-orange-500 p-2 font-mono font-bold text-white" href="/">skeleton-svelte</a>
-			</header>
-			<hr class="hr" />
-			<Switch
-				name="mode"
-				controlActive="bg-surface-200"
-				checked={modeState}
-				onCheckedChange={(e) => {
-					modeState = e.checked;
-					toggleMode();
-				}}
-			>
-				{#snippet inactiveChild()}<IconMoon size="14" />{/snippet}
-				{#snippet activeChild()}<IconSun size="14" />{/snippet}
-				<p>Toggle Mode</p>
-			</Switch>
-			<hr class="hr" />
-			<!-- Components -->
-			<div class="space-y-8">
-				<div class="font-bold">Components</div>
-				<nav class="text-sm flex flex-col gap-2">
-					<a class="anchor" href="/components/accordions">Accordions</a>
-					<a class="anchor" href="/components/app-bars">App Bars</a>
-					<a class="anchor" href="/components/avatars">Avatars</a>
-					<a class="anchor" href="/components/file-upload">File Upload</a>
-					<a class="anchor" href="/components/navigation">Navigation</a>
-					<a class="anchor" href="/components/pagination">Pagination</a>
-					<a class="anchor" href="/components/progress">Progress</a>
-					<a class="anchor" href="/components/progress-rings">Progress Rings</a>
-					<a class="anchor" href="/components/ratings">Ratings</a>
-					<a class="anchor" href="/components/segment-control">Segment Control</a>
-					<a class="anchor" href="/components/slider">Slider</a>
-					<a class="anchor" href="/components/switch">Switch</a>
-					<a class="anchor" href="/components/tabs">Tabs</a>
-					<a class="anchor" href="/components/tags-input">Tags Input</a>
-					<a class="anchor" href="/components/toast">Toast</a>
-				</nav>
-			</div>
-			<!-- Svelte-Only -->
-			<div class="space-y-8">
-				<div class="font-bold">Svelte-Only</div>
-				<nav class="text-sm flex flex-col gap-2">
-					<a class="anchor" href="/components/popover">Popover</a>
-					<a class="anchor" href="/components/tooltip">Tooltip</a>
-					<a class="anchor" href="/components/combobox">Combobox</a>
-					<a class="anchor" href="/components/modal">Modal</a>
-				</nav>
-			</div>
+<Toaster {toaster}></Toaster>
+
+<div class="grid h-screen grid-cols-[320px_minmax(0,_1fr)]">
+	<!-- Nav -->
+	<div class="bg-surface-100-900 space-y-8 overflow-y-auto p-8">
+		<header>
+			<a class="text-sm bg-orange-500 p-2 font-mono font-bold text-white" href="/">skeleton-svelte</a>
+		</header>
+		<hr class="hr" />
+		<Switch
+			name="mode"
+			controlActive="bg-surface-200"
+			checked={modeState}
+			onCheckedChange={(e) => {
+				modeState = e.checked;
+				toggleMode();
+			}}
+		>
+			{#snippet inactiveChild()}<IconMoon size="14" />{/snippet}
+			{#snippet activeChild()}<IconSun size="14" />{/snippet}
+			<p>Toggle Mode</p>
+		</Switch>
+		<hr class="hr" />
+		<!-- Components -->
+		<div class="space-y-8">
+			<div class="font-bold">Components</div>
+			<nav class="text-sm flex flex-col gap-2">
+				<a class="anchor" href="/components/accordions">Accordions</a>
+				<a class="anchor" href="/components/app-bars">App Bars</a>
+				<a class="anchor" href="/components/avatars">Avatars</a>
+				<a class="anchor" href="/components/file-upload">File Upload</a>
+				<a class="anchor" href="/components/navigation">Navigation</a>
+				<a class="anchor" href="/components/pagination">Pagination</a>
+				<a class="anchor" href="/components/progress">Progress</a>
+				<a class="anchor" href="/components/progress-rings">Progress Rings</a>
+				<a class="anchor" href="/components/ratings">Ratings</a>
+				<a class="anchor" href="/components/segment-control">Segment Control</a>
+				<a class="anchor" href="/components/slider">Slider</a>
+				<a class="anchor" href="/components/switch">Switch</a>
+				<a class="anchor" href="/components/tabs">Tabs</a>
+				<a class="anchor" href="/components/tags-input">Tags Input</a>
+				<a class="anchor" href="/components/toast">Toast</a>
+			</nav>
 		</div>
-		<!-- Page -->
-		<main class="space-y-8 overflow-y-auto p-8">
-			{@render children?.()}
-		</main>
+		<!-- Svelte-Only -->
+		<div class="space-y-8">
+			<div class="font-bold">Svelte-Only</div>
+			<nav class="text-sm flex flex-col gap-2">
+				<a class="anchor" href="/components/popover">Popover</a>
+				<a class="anchor" href="/components/tooltip">Tooltip</a>
+				<a class="anchor" href="/components/combobox">Combobox</a>
+				<a class="anchor" href="/components/modal">Modal</a>
+			</nav>
+		</div>
 	</div>
-</ToastProvider>
+	<!-- Page -->
+	<main class="space-y-8 overflow-y-auto p-8">
+		{@render children?.()}
+	</main>
+</div>
